fix(symthaea): standardize CoreAffect to f32, add 15 ContinuousHV property tests
- Fix PhysiologicalState f64→f32 mismatch (affective_consciousness.rs)
- Add f64→f32 casts in InteroceptiveState::to_physiological() (embodied_cognition.rs)
- Add 15 ContinuousHV property tests: self-similarity, near-orthogonality,
  bind commutativity/dissimilarity/similarity-preservation, bundle component
  preservation/identity, similarity bounds, normalize unit norm, permute
  cyclic/inverse/changes-vector, negation, zero vector (test_hdc_properties.rs)
- All 27 HDC property tests pass (12 existing BinaryHV + 15 new ContinuousHV)
- 0 clippy warnings

+++ symthaea/src/consciousness/embodied_cognition.rs
@@ -389,11 +389,11 @@ impl InteroceptiveState {
     /// Convert to physiological state (bridge to #88)
     pub fn to_physiological(&self) -> PhysiologicalState {
         PhysiologicalState {
-            heart_rate_delta: self.heart_rate - 0.3, // Deviation from baseline
-            skin_conductance: self.visceral_arousal * 0.5,
-            muscle_tension: self.fatigue * 0.3,
-            respiration_delta: self.breathing_rate - 0.3,
-            gut_feeling: self.gut_state,
+            heart_rate_delta: (self.heart_rate - 0.3) as f32,
+            skin_conductance: (self.visceral_arousal * 0.5) as f32,
+            muscle_tension: (self.fatigue * 0.3) as f32,
+            respiration_delta: (self.breathing_rate - 0.3) as f32,
+            gut_feeling: self.gut_state as f32,
         }
     }
